Setup dependencies for subproject building
We need to define dependency information so that other projects
can build libhei as a subproject.

+++ src/isolator/hei_isolator.cpp
@@ -19,8 +19,8 @@ ReturnCode Isolator::initialize(void* i_buffer, size_t i_bufferSize,
     ReturnCode rc;
 
     // BEGIN temporary code
-    HEI_INF("Isolator::initialize(%p,%lu,%d)", i_buffer, i_bufferSize,
-            i_forceInit);
+    HEI_INF("Isolator::initialize(%p,%" PRIu64 ",%d)", i_buffer,
+            (uint64_t)i_bufferSize, i_forceInit);
 
     auto& scom_fw    = Flyweight<ScomRegister>::getSingleton();
     auto& idScom_fw  = Flyweight<IdScomRegister>::getSingleton();
